+++ src/notation/internal/notationactioncontroller.cpp
@@ -439,7 +439,7 @@ void NotationActionController::init()
         }
     }
 
-    for (int i = 0; i < Ms::VOICES; ++i) {
+    for (size_t i = 0; i < Ms::VOICES; ++i) {
         registerAction("voice-" + std::to_string(i + 1), [this, i]() { changeVoice(i); });
     }
